+++ src/tests/data/bn/invmod.vec
@@ -487,3 +487,44 @@ Input = 657900513264442578215729259525804042708991731028255426638688946278845827
 Modulus = 1460598742727828605836860681156746006166272494387112693708892744343548055514972863039476201659720247220048664250204314648520085411164461712526657028588699682983099362771
 Output = 370134468809235355809931021496418560257927761651782631431723120822241707286126622619231228275210678415132318220494260963802381448709723310690465171935975287188943190781
 
+# Moduli power of 2
+Input = 3
+Modulus = 0x10
+Output  = 11
+
+Input = 11
+Modulus = 0x100
+Output  = 163
+
+Input = 163
+Modulus = 0x1000
+Output  = 779
+
+Input = 2731
+Modulus = 0x10000
+Output  = 40963
+
+Input = 40963
+Modulus = 0x100000
+Output  = 461483
+
+Input = 461483
+Modulus = 0x1000000
+Output  = 12623875
+
+Input = 12623875
+Modulus = 0x10000000
+Output  = 50793131
+
+Input = 50793131
+Modulus = 0x100000000
+Output  = 2965413891
+
+Input = 2965413891
+Modulus = 0x1000000000
+Output  = 43000466091
+
+Input = 43000466091
+Modulus = 0x10000000000
+Output  = 140404367363
+
